Implementing update user method and improving some points

+++ src/bll/users.js
@@ -54,10 +54,12 @@ function getAllUsers (callback) {
  * @callback Callback Request callback.
  */
 function getUserById (userId, callback) {
-    let query = "SELECT * FROM user WHERE id = ?;";
-    let params = [];
-    params.push(userId);
-    executeQuery(query, params, callback);
+    if(userId) {
+        let query = "SELECT * FROM user WHERE id = ?;";
+        let params = [];
+        params.push(userId);
+        executeQuery(query, params, callback);
+    }    
 }
 
 /**
@@ -66,27 +68,51 @@ function getUserById (userId, callback) {
  * @callback Callback Request callback.
  */
 function deleteUser (userId, callback) {
-    // Search for the user in DB.
-    let select = "SELECT * FROM user WHERE id = ?;"
-    let params = [];
-    params.push(userId);
-    executeQuery(select, params, (error, result) => {
-        if(error) callback(error);
-        // If the user was found, execute the delete.
-        if(result && result.length) {
-            let query = "DELETE FROM user WHERE id = ?;";    
-            executeQuery(query, params, callback);
-        } else {
-            callback(new Error('User not found'));
-        }
-    });
+    if(userId) {
+        // Search for the user in DB.
+        let select = "SELECT * FROM user WHERE id = ?;"
+        let params = [];
+        params.push(userId);
+        executeQuery(select, params, (error, result) => {
+            if(error) callback(error);
+            // If the user was found, execute the delete.
+            if(result && result.length) {
+                let query = "DELETE FROM user WHERE id = ?;";    
+                executeQuery(query, params, callback);
+            } else {
+                callback(new Error('User not found'));
+            }
+        });
+    }    
+}
+
+function updateUser (user, callback) {
+    if(user) {
+        // Search for the user in DB.
+        let select = "SELECT * FROM user WHERE id = ?;"  
+        executeQuery(select, [user.id], (error, result) => {
+            if(error) callback(error);
+            // If the user was found, execute the update.
+            if(result && result.length) {
+                let query = "UPDATE user SET name = ?, email = ? WHERE id = ?;";
+                let updtParams = [];
+                updtParams.push(user.name);
+                updtParams.push(user.email);
+                updtParams.push(user.id);
+                executeQuery(query, updtParams, callback);
+            } else {
+                callback(new Error('User not found'));
+            }
+        });
+    }
 }
 
 const userService = {
     insertUser: insertUser,
     getAllUsers: getAllUsers,
     getUserById: getUserById,
-    deleteUser: deleteUser
+    deleteUser: deleteUser,
+    updateUser: updateUser
 }
 
 module.exports = userService;
